time: timezone should be set by backend based on provided region and city
Currently front-end need to know the GNU timezone name (e.g. `CET-1CEST,M3.5.0,M10.5.0/3`)
Which is bit harder to get and it makes no sense to have it
as a part of an API call. It can be derived from Region/Country

This commit alters backend to ignore provided `timezone` attribute,
makes it optional to `update_settings` and sets timezone properly