updater: Adjust handling of "reboot is required" from updater supervisor
Changed in updater supervisor made validation of API call `updater->get_settings`
fail on `reboot` attribute in JSON reply.

Which in turn makes updates page in reForis unusable, because error is shown
instead of update summary. That way it is not possible to approve update request.

For relevant changes in updater-supervisor see:
https://gitlab.nic.cz/turris/updater/supervisor/-/commit/aecdf2c4b003d031a5a84116118053daa30d50f6

Closes: #241